+++ tests/src/tests/streamswap_tests/shares.rs
@@ -16,6 +16,7 @@ mod shares {
             Timestamp::from_seconds(0),
             Timestamp::from_seconds(100),
             Timestamp::from_seconds(0),
+            None,
         );
 
         // add new shares
